+++ example/lib/pages/controls_configuration_page.dart
@@ -23,6 +23,7 @@ class _ControlsConfigurationPageState extends State<ControlsConfigurationPage> {
       enableSkips: false,
       enableFullscreen: false,
       controlBarHeight: 60,
+      loadingColor: Colors.red,
     );
 
     BetterPlayerConfiguration betterPlayerConfiguration =
@@ -31,7 +32,7 @@ class _ControlsConfigurationPageState extends State<ControlsConfigurationPage> {
             fit: BoxFit.contain,
             controlsConfiguration: controlsConfiguration);
     BetterPlayerDataSource dataSource = BetterPlayerDataSource(
-        BetterPlayerDataSourceType.network, Constants.forBiggerBlazesUrl);
+        BetterPlayerDataSourceType.network, Constants.elephantDreamVideoUrl);
     _betterPlayerController = BetterPlayerController(betterPlayerConfiguration);
     _betterPlayerController.setupDataSource(dataSource);
     super.initState();