+++ example/lib/pages/normal_player_page.dart
@@ -19,7 +19,8 @@ class _NormalPlayerPageState extends State<NormalPlayerPage> {
       fit: BoxFit.contain,
     );
     BetterPlayerDataSource dataSource = BetterPlayerDataSource(
-        BetterPlayerDataSourceType.network, "https://v5.szjal.cn/20200829/qByymCNH/index.m3u8");
+        BetterPlayerDataSourceType.network,
+        "https://v5.szjal.cn/20200829/qByymCNH/index.m3u8");
     _betterPlayerController = BetterPlayerController(betterPlayerConfiguration);
     _betterPlayerController.setupDataSource(dataSource);