+++ lib/src/core/better_player_utils.dart
@@ -61,6 +61,7 @@ class BetterPlayerUtils {
 
   static void log(String logMessage) {
     if (!kReleaseMode) {
+      // ignore: avoid_print
       print(logMessage);
     }
   }