+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.0.40
 * Exposed VideoPlayerValue in export
 * Fixed log issue
+* Added loadingColor and loadingWidget in BetterPlayerControlsConfiguration
 
 ## 0.0.39
 * Added lint library for dart analysis